+++ src/embedded_deps/standard_pbr_frag.glsl
@@ -9,10 +9,7 @@ in vec3 vNormal;
 #endif
 in vec2 vUV;
 in vec3 vWsPos;
-in vec3 vLightDir[SI_LIGHTS];
 
-uniform vec3 g_ambientLight;
-uniform vec4 g_lightColorRange[SI_LIGHTS];
 uniform vec4 color;
 uniform vec4 metallicRoughness;
 uniform vec4 g_cameraPos;
@@ -37,6 +34,7 @@ in vec4 vColor;
 #endif
 
 #pragma include "normalmap_incl.glsl"
+#pragma include "light_incl.glsl"
 #pragma include "sre_utils_incl.glsl"
 
 
@@ -147,7 +145,13 @@ void main(void)
     vec3 n = getNormal();                             // Normal at surface point
     vec3 v = normalize(g_cameraPos.xyz - vWsPos.xyz); // Vector from surface point to camera
     for (int i=0;i<SI_LIGHTS;i++) {
-        vec3 l = normalize(vLightDir[i]);                 // Vector from surface point to light
+        float attenuation = 0.0;
+        vec3 l = vec3(0.0,0.0,0.0);
+        lightDirectionAndAttenuation(g_lightPosType[i], g_lightColorRange[i].w, vWsPos, l, attenuation);
+        if (attenuation <= 0.0){
+            continue;
+        }
+
         vec3 h = normalize(l+v);                          // Half vector between both l and v
         vec3 reflection = -normalize(reflect(v, n));
 
@@ -180,7 +184,7 @@ void main(void)
         // Calculation of analytical lighting contribution
         vec3 diffuseContrib = (1.0 - F) * diffuse(pbrInputs);
         vec3 specContrib = F * G * D / (4.0 * NdotL * NdotV);
-        color += NdotL * g_lightColorRange[i].xyz * (diffuseContrib + specContrib);
+        color += attenuation * NdotL * g_lightColorRange[i].xyz * (diffuseContrib + specContrib);
     }
 
     // Apply optional PBR terms for additional (optional) shading