+++ version-history.md
@@ -1,5 +1,7 @@
 ## Version history
 
+ * 1.0.2 Add Phong shader. Improve light (PBR and Blinn-Phong).
+ * 1.0.1 Improve specular highlight for Blinn-Phong. Add attenuation to PBR.
  * 1.0.0 Physically based rendering. Profiler renamed to Inspector. Added shader specialization constants. Trimmed sprites.
  * 0.9.23 Add support for geometry and tesselation shaders. Fix MSVC compile error.
  * 0.9.22 OpenVR support (experimental). Queue draw calls (to better support simultaneous renderpasses). Make max scenelight count depends on Renderer::maxSceneLights. Optimize and clarify sprite batch order.